Experimental i2c update for Notecard ESP compatbility
The Notecard ESP's slave mode receive function has notably higher
latency than the STM32, particularly so when the ESP cores are busy,
e.g. when syncing.

Experimentally it was found that this could lead to
- Receive buffer overflow and data loss
- Failure to prefill the Notecard transmit buffer in time for the read
  request
- Out of sync condition where the read transaction reads data queued by a
  prior write transaction.

In order to mitigate these problems much larger delays are introduced
between write transactions, and between the write and subsequent read
transaction. Additionally if an error occurs in a read transaction the
delays are significantly increased, in case the Notecard is syncing.
The delays are reduced back to normal on successive successful read
transactions.