fix: reduce current usage on STM32-based hosts when using TXNs
When using transactions (for example, when talking to a NOTE-ESP) this library must initialize and uninitialize GPIOs for RTX and CTX.

For system current usage reasons - particularly, if the sketch is using the Low Power library - it is important that we don't put IO pins with no pullups or pulldowns into an ambiguous GPIO state that will cause excessive current draw.

This PR leverages the STM32's ability to actually put a pin into a floating state with no leakage.